requirements.txt
# Core dependencies
python-binance==1.0.19
pandas==2.1.4
numpy==1.26.2
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0

# Data processing
ta==0.11.0
ta-lib==0.4.28

# Backend
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# WebSocket
websocket-client==1.7.0
websockets==12.0

# Queue/Task processing
redis==5.0.1
celery==5.3.4

# Logging & Metrics
structlog==24.1.0
prometheus-client==0.19.0

# Strategy validation
pydantic==2.5.3

# Risk management
httpx==0.26.0  # Already included but needed for Slack alerts

# Observability
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-sqlalchemy
opentelemetry-exporter-otlp-proto-grpc
psutil

# Optional
aiofiles==23.2.1

